# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API",
        "description": "Lorem Ipsum",
        "contact": {
            "email": "hi.wasissubekti02@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/our-services": {
            "get": {
                "tags": [
                    "Our Services"
                ],
                "summary": "ambil semua Our Services",
                "description": "Mengambil Data Semua Our Services",
                "operationId": "our_services",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "success get all our_services",
                                    "data": [
                                        {
                                            "id": 15,
                                            "image": "1801550969236851.jpg",
                                            "title": "ok",
                                            "description": "Sip",
                                            "created_at": "2024-06-10T02:36:13.000000Z",
                                            "updated_at": "2024-06-11T08:01:32.000000Z"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Our Services"
                ],
                "summary": "Create new Our Services",
                "description": "Create a new Our Services",
                "operationId": "create_our_services",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "description": "The image of Our Services",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "description": {
                                        "description": "The description of Our Services",
                                        "type": "string"
                                    },
                                    "title": {
                                        "description": "The title of Our Services",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": true,
                                    "message": "success post Our Services",
                                    "data": {
                                        "id": 15,
                                        "image": "1801550969236851.jpg",
                                        "title": "ok",
                                        "description": "Sip",
                                        "created_at": "2024-06-10T02:36:13.000000Z",
                                        "updated_at": "2024-06-11T08:01:32.000000Z"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": false,
                                    "message": "fail post Our Services",
                                    "error": "Validation error message"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/our-services/{id}": {
            "get": {
                "tags": [
                    "Our Services"
                ],
                "summary": "Get single Our Services",
                "description": "Retrieve a single Our Services by ID",
                "operationId": "single_our_services",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the Our Services",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": true,
                                    "message": "success get Our Services",
                                    "data": {
                                        "id": 15,
                                        "image": "1801550969236851.jpg",
                                        "title": "ok",
                                        "description": "Sip",
                                        "created_at": "2024-06-10T02:36:13.000000Z",
                                        "updated_at": "2024-06-11T08:01:32.000000Z"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Our Services not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": false,
                                    "message": "fail get Our Services",
                                    "error": "No query results for model [App\\\\Models\\\\OurServices] 10"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Our Services"
                ],
                "summary": "Update Our Services",
                "description": "Update Our Services by ID",
                "operationId": "update_our_services",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of Our Services",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "description": "The image of Our Services",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "description": {
                                        "description": "The description of Our Services",
                                        "type": "string"
                                    },
                                    "title": {
                                        "description": "The title of Our Services",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": true,
                                    "message": "success update Our Services",
                                    "data": {
                                        "id": 15,
                                        "image": "1801550969236851.jpg",
                                        "title": "ok",
                                        "description": "Sip",
                                        "created_at": "2024-06-10T02:36:13.000000Z",
                                        "updated_at": "2024-06-11T08:01:32.000000Z"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Our Services not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": false,
                                    "message": "fail update Our Services",
                                    "error": "No query results for model [App\\\\Models\\\\OurServices] 10"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/our-service/{id}": {
            "delete": {
                "tags": [
                    "Our Services"
                ],
                "summary": "Delete Our Services",
                "description": "Delete Our Services by ID",
                "operationId": "delete_our_services",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the Our Services",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": true,
                                    "message": "success delete Our Services"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Our Services not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": false,
                                    "message": "fail delete Our Services",
                                    "error": "No query results for model [App\\\\Models\\\\OurServices] 10"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}